PIN-3695: Minor code refactor
